+++ packages/contracts/test/voting/multisig.ts
@@ -99,6 +99,20 @@ describe('Multisig', function () {
     );
   });
 
+  function addresslist(length: number): string[] {
+    let addresses: string[] = [];
+
+    for (let i = 0; i < length; i++) {
+      const addr = signers[i].address;
+      addresses.push(addr);
+    }
+    return addresses;
+  }
+
+  async function timestampIn(durationInSec: number): Promise<number> {
+    return (await ethers.provider.getBlock('latest')).timestamp + durationInSec;
+  }
+
   describe('initialize:', async () => {
     it('reverts if trying to re-initialize', async () => {
       await multisig.initialize(
@@ -302,7 +316,14 @@ describe('Multisig', function () {
       expect(await multisig.proposalCount()).to.equal(0);
 
       await expect(
-        multisig.createProposal(dummyMetadata, dummyActions, false, false)
+        multisig.createProposal(
+          dummyMetadata,
+          dummyActions,
+          false,
+          false,
+          0,
+          await timestampIn(1000)
+        )
       ).to.not.be.reverted;
 
       expect(await multisig.proposalCount()).to.equal(1);
@@ -320,18 +341,29 @@ describe('Multisig', function () {
         dummyMetadata,
         dummyActions,
         false,
-        false
+        false,
+        0,
+        await timestampIn(1000)
       );
       // create a new proposal for the proposalCounter to be incremented
       await expect(
-        multisig.createProposal(dummyMetadata, dummyActions, false, false)
+        multisig.createProposal(
+          dummyMetadata,
+          dummyActions,
+          false,
+          false,
+          0,
+          await timestampIn(1000)
+        )
       ).to.not.be.reverted;
 
       let proposalId1 = await multisig.callStatic.createProposal(
         dummyMetadata,
         dummyActions,
         false,
-        false
+        false,
+        0,
+        await timestampIn(1000)
       );
 
       expect(proposalId0).to.equal(0); // To be removed when proposal ID is generated as a hash.
@@ -347,13 +379,22 @@ describe('Multisig', function () {
         multisigSettings
       );
 
+      let startDate = await timestampIn(1000);
+      let endDate = await timestampIn(5000);
       await expect(
         multisig
           .connect(signers[0])
-          .createProposal(dummyMetadata, [], false, false)
+          .createProposal(dummyMetadata, [], false, false, startDate, endDate)
       )
         .to.emit(multisig, PROPOSAL_EVENTS.PROPOSAL_CREATED)
-        .withArgs(id, signers[0].address, dummyMetadata, []);
+        .withArgs(
+          id,
+          signers[0].address,
+          startDate,
+          endDate,
+          dummyMetadata,
+          []
+        );
     });
 
     context('`onlyListed` is set to `false`:', async () => {
@@ -368,13 +409,23 @@ describe('Multisig', function () {
       });
 
       it('creates a proposal when unlisted accounts are allowed', async () => {
+        let startDate = await timestampIn(1000);
+        let endDate = await timestampIn(5000);
+
         await expect(
           multisig
             .connect(signers[1]) // not listed
-            .createProposal(dummyMetadata, [], false, false)
+            .createProposal(dummyMetadata, [], false, false, startDate, endDate)
         )
           .to.emit(multisig, PROPOSAL_EVENTS.PROPOSAL_CREATED)
-          .withArgs(id, signers[1].address, dummyMetadata, []);
+          .withArgs(
+            id,
+            signers[1].address,
+            startDate,
+            endDate,
+            dummyMetadata,
+            []
+          );
       });
     });
 
@@ -393,7 +444,14 @@ describe('Multisig', function () {
         await expect(
           multisig
             .connect(signers[1]) // not listed
-            .createProposal(dummyMetadata, [], false, false)
+            .createProposal(
+              dummyMetadata,
+              [],
+              false,
+              false,
+              0,
+              await timestampIn(1000)
+            )
         )
           .to.be.revertedWithCustomError(multisig, 'ProposalCreationForbidden')
           .withArgs(signers[1].address);
@@ -401,14 +459,42 @@ describe('Multisig', function () {
         await expect(
           multisig
             .connect(signers[0])
-            .createProposal(dummyMetadata, [], false, false)
+            .createProposal(
+              dummyMetadata,
+              [],
+              false,
+              false,
+              0,
+              await timestampIn(1000)
+            )
         ).to.not.be.reverted;
       });
 
       it('creates a proposal successfully and does not approve if not specified', async () => {
-        await expect(multisig.createProposal(dummyMetadata, [], false, false))
+        let startDate = await timestampIn(1000);
+        let endDate = await timestampIn(5000);
+
+        await ethers.provider.send('evm_setNextBlockTimestamp', [startDate]);
+
+        await expect(
+          multisig.createProposal(
+            dummyMetadata,
+            [],
+            false,
+            false,
+            startDate,
+            endDate
+          )
+        )
           .to.emit(multisig, PROPOSAL_EVENTS.PROPOSAL_CREATED)
-          .withArgs(id, signers[0].address, dummyMetadata, []);
+          .withArgs(
+            id,
+            signers[0].address,
+            startDate,
+            endDate,
+            dummyMetadata,
+            []
+          );
 
         const block = await ethers.provider.getBlock('latest');
 
@@ -418,6 +504,8 @@ describe('Multisig', function () {
         expect(proposal.parameters.minApprovals).to.equal(
           multisigSettings.minApprovals
         );
+        expect(proposal.parameters.startDate).to.equal(startDate);
+        expect(proposal.parameters.endDate).to.equal(endDate);
         expect(proposal.tally.approvals).to.equal(0);
         expect(proposal.tally.addresslistLength).to.equal(1);
         expect(proposal.actions.length).to.equal(0);
@@ -427,31 +515,60 @@ describe('Multisig', function () {
       });
 
       it('creates a proposal successfully and approves if specified', async () => {
-        await expect(multisig.createProposal(dummyMetadata, [], true, false))
+        let startDate = await timestampIn(3000);
+        let endDate = await timestampIn(5000);
+
+        await ethers.provider.send('evm_setNextBlockTimestamp', [startDate]);
+
+        await expect(
+          multisig.createProposal(dummyMetadata, [], true, false, 0, endDate)
+        )
           .to.emit(multisig, PROPOSAL_EVENTS.PROPOSAL_CREATED)
-          .withArgs(id, signers[0].address, dummyMetadata, [])
+          .withArgs(
+            id,
+            signers[0].address,
+            startDate,
+            endDate,
+            dummyMetadata,
+            []
+          )
           .to.emit(multisig, MULTISIG_EVENTS.APPROVED)
           .withArgs(id, signers[0].address);
 
         const block = await ethers.provider.getBlock('latest');
+
         const proposal = await multisig.getProposal(id);
-        expect(proposal.open).to.equal(true);
         expect(proposal.executed).to.equal(false);
         expect(proposal.parameters.snapshotBlock).to.equal(block.number - 1);
         expect(proposal.parameters.minApprovals).to.equal(
           multisigSettings.minApprovals
         );
-
+        expect(proposal.parameters.startDate).to.equal(startDate);
+        expect(proposal.parameters.endDate).to.equal(endDate);
         expect(proposal.tally.approvals).to.equal(1);
       });
 
       it('increases the proposal count', async () => {
         expect(await multisig.proposalCount()).to.equal(0);
 
-        await multisig.createProposal(dummyMetadata, dummyActions, true, false);
+        await multisig.createProposal(
+          dummyMetadata,
+          dummyActions,
+          true,
+          false,
+          0,
+          await timestampIn(5000)
+        );
         expect(await multisig.proposalCount()).to.equal(1);
 
-        await multisig.createProposal(dummyMetadata, dummyActions, true, false);
+        await multisig.createProposal(
+          dummyMetadata,
+          dummyActions,
+          true,
+          false,
+          0,
+          await timestampIn(5000)
+        );
         expect(await multisig.proposalCount()).to.equal(2);
       });
     });
@@ -466,16 +583,14 @@ describe('Multisig', function () {
         multisigSettings
       );
 
-      expect(
-        (
-          await multisig.createProposal(
-            dummyMetadata,
-            dummyActions,
-            false,
-            false
-          )
-        ).value
-      ).to.equal(id);
+      await multisig.createProposal(
+        dummyMetadata,
+        dummyActions,
+        false,
+        false,
+        0,
+        await timestampIn(1000)
+      );
     });
 
     describe('canApprove:', async () => {
@@ -501,6 +616,42 @@ describe('Multisig', function () {
 
       it('returns `true` if the approver is listed', async () => {
         expect(await multisig.canApprove(id, signers[0].address)).to.be.true;
+      });
+
+      it("returns `false` if the proposal hasn't started yet", async () => {
+        await multisig.createProposal(
+          dummyMetadata,
+          dummyActions,
+          false,
+          false,
+          await timestampIn(2000),
+          await timestampIn(5000)
+        );
+
+        expect(await multisig.canApprove(1, signers[0].address)).to.be.false;
+
+        await ethers.provider.send('evm_increaseTime', [2000]);
+        await ethers.provider.send('evm_mine');
+
+        expect(await multisig.canApprove(1, signers[0].address)).to.be.true;
+      });
+
+      it('returns `false` if the proposal has ended', async () => {
+        await multisig.createProposal(
+          dummyMetadata,
+          dummyActions,
+          false,
+          false,
+          0,
+          await timestampIn(5000)
+        );
+
+        expect(await multisig.canApprove(1, signers[0].address)).to.be.true;
+
+        await ethers.provider.send('evm_increaseTime', [5000]);
+        await ethers.provider.send('evm_mine');
+
+        expect(await multisig.canApprove(1, signers[0].address)).to.be.false;
       });
     });
 
@@ -533,6 +684,42 @@ describe('Multisig', function () {
 
         expect((await multisig.getProposal(id)).tally.approvals).to.equal(1);
       });
+
+      it("reverts if the proposal hasn't started yet", async () => {
+        await multisig.createProposal(
+          dummyMetadata,
+          dummyActions,
+          false,
+          false,
+          await timestampIn(5000),
+          await timestampIn(10000)
+        );
+
+        await expect(multisig.approve(1, false)).to.be.reverted;
+
+        await ethers.provider.send('evm_increaseTime', [7000]);
+
+        await expect(multisig.approve(1, false)).to.not.be.reverted;
+      });
+
+      it('reverts if the proposal has ended', async () => {
+        await multisig.createProposal(
+          dummyMetadata,
+          dummyActions,
+          false,
+          false,
+          0,
+          await timestampIn(5000)
+        );
+        await ethers.provider.send('evm_increaseTime', [2000]);
+
+        await expect(multisig.connect(signers[1]).approve(1, false)).to.not.be
+          .reverted;
+
+        await ethers.provider.send('evm_increaseTime', [10000]);
+
+        await expect(multisig.approve(1, false)).to.be.reverted;
+      });
     });
 
     describe('canExecute:', async () => {
@@ -559,6 +746,50 @@ describe('Multisig', function () {
         expect((await multisig.getProposal(id)).executed).to.be.false;
 
         expect(await multisig.canExecute(id)).to.be.true;
+      });
+
+      it("returns `false` if the proposal hasn't started yet", async () => {
+        await multisig.createProposal(
+          dummyMetadata,
+          dummyActions,
+          false,
+          false,
+          await timestampIn(1000),
+          await timestampIn(10000)
+        );
+
+        expect(await multisig.canExecute(1)).to.be.false;
+
+        await ethers.provider.send('evm_increaseTime', [2000]);
+        await multisig.connect(signers[0]).approve(1, false);
+        await multisig.connect(signers[1]).approve(1, false);
+        await multisig.connect(signers[2]).approve(1, false);
+
+        expect(await multisig.canExecute(1)).to.be.true;
+      });
+
+      it('returns `false` if the proposal has ended', async () => {
+        await multisig.createProposal(
+          dummyMetadata,
+          dummyActions,
+          false,
+          false,
+          0,
+          await timestampIn(5000)
+        );
+
+        await ethers.provider.send('evm_increaseTime', [2000]);
+
+        await multisig.connect(signers[0]).approve(1, false);
+        await multisig.connect(signers[1]).approve(1, false);
+        await multisig.connect(signers[2]).approve(1, false);
+
+        expect(await multisig.canExecute(1)).to.be.true;
+
+        await ethers.provider.send('evm_increaseTime', [5000]);
+        await ethers.provider.send('evm_mine');
+
+        expect(await multisig.canExecute(1)).to.be.false;
       });
     });
 
@@ -663,6 +894,44 @@ describe('Multisig', function () {
           .to.emit(dao, DAO_EVENTS.EXECUTED)
           .to.emit(multisig, PROPOSAL_EVENTS.PROPOSAL_EXECUTED)
           .to.emit(multisig, MULTISIG_EVENTS.APPROVED);
+      });
+
+      it("reverts if the proposal hasn't started yet", async () => {
+        await multisig.createProposal(
+          dummyMetadata,
+          dummyActions,
+          false,
+          false,
+          await timestampIn(1000),
+          await timestampIn(5000)
+        );
+
+        await expect(multisig.execute(1)).to.be.reverted;
+
+        await ethers.provider.send('evm_increaseTime', [2000]);
+
+        await multisig.connect(signers[0]).approve(1, false);
+        await multisig.connect(signers[1]).approve(1, false);
+        await multisig.connect(signers[2]).approve(1, false);
+
+        await expect(multisig.execute(1)).to.not.be.reverted;
+      });
+
+      it('reverts if the proposal has ended', async () => {
+        await multisig.createProposal(
+          dummyMetadata,
+          dummyActions,
+          false,
+          false,
+          0,
+          await timestampIn(5000)
+        );
+        await multisig.connect(signers[0]).approve(1, false);
+        await multisig.connect(signers[1]).approve(1, false);
+        await multisig.connect(signers[2]).approve(1, false);
+
+        await ethers.provider.send('evm_increaseTime', [10000]);
+        await expect(multisig.connect(signers[1]).execute(1)).to.be.reverted;
       });
     });
   });