+++ packages/subgraph/CHANGELOG.md
@@ -7,8 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [UPCOMING]
 
+### Added
+- Adds `startDate` and `endDate` fields to Multisig proposals
+
 ### Changed
 - Ignores `None` votes from addresslist voting and token voting
+
+### Removed
+- Removes `open` and `executable` fields from Multisig proposals 
 
 ## 0.5.0-alpha